feat: authorize checklist item

+++ app/Http/Controllers/Item/DeleteChecklist.php
@@ -4,7 +4,10 @@ namespace App\Http\Controllers\Item;
 
 use App\Http\Controllers\Controller;
 use App\Models\ChecklistItem;
+use App\Models\Item;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\DB;
 
 class DeleteChecklist extends Controller
 {
@@ -20,6 +23,17 @@ class DeleteChecklist extends Controller
             );
         }
 
+        $item = Item::find($checklist->item_id);
+        if (!$item || !$this->checkPermission($item->container->board_id)) {
+            return response(
+                [
+                    'message' =>
+                        'You do not have permission to manage checklist in item',
+                ],
+                403
+            );
+        }
+
         $checklist->delete();
 
         return response(
@@ -28,5 +42,24 @@ class DeleteChecklist extends Controller
             ],
             200
         );
+    }
+
+    private function checkPermission($board_id)
+    {
+        $user_id = Auth::id();
+
+        $checklist_management = DB::table('user_in_board')
+            ->join(
+                'board_roles',
+                'user_in_board.board_role_id',
+                '=',
+                'board_roles.id'
+            )
+            ->where('user_in_board.user_id', $user_id)
+            ->where('user_in_board.board_id', $board_id)
+            ->select('board_roles.checklist_management')
+            ->first()->checklist_management;
+
+        return $checklist_management;
     }
 }
